move identity serializer and mocks to protoutil
