+++ common/deliverclient/blocksprovider/bft_deliverer_test.go
@@ -31,7 +31,7 @@ import (
 	"github.com/hyperledger/fabric-x-common/common/deliverclient/orderers"
 	"github.com/hyperledger/fabric-x-common/common/util"
 	"github.com/hyperledger/fabric-x-common/protoutil"
-	"github.com/hyperledger/fabric-x-common/tools/pkg/identity/mocks"
+	"github.com/hyperledger/fabric-x-common/protoutil/identity/mocks"
 )
 
 type bftDelivererTestSetup struct {
