+++ common/deliverclient/blocksprovider/deliverer.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/hyperledger/fabric-x-common/common/deliverclient/orderers"
 	"github.com/hyperledger/fabric-x-common/common/util"
-	"github.com/hyperledger/fabric-x-common/tools/pkg/identity"
+	"github.com/hyperledger/fabric-x-common/protoutil/identity"
 )
 
 // LedgerInfo an adapter to provide the interface to query
