+++ common/deliverclient/blocksprovider/deliverer_test.go
@@ -30,8 +30,8 @@ import (
 	"github.com/hyperledger/fabric-x-common/common/util"
 	"github.com/hyperledger/fabric-x-common/core/config/configtest"
 	"github.com/hyperledger/fabric-x-common/protoutil"
+	"github.com/hyperledger/fabric-x-common/protoutil/identity/mocks"
 	"github.com/hyperledger/fabric-x-common/tools/configtxgen"
-	"github.com/hyperledger/fabric-x-common/tools/pkg/identity/mocks"
 	"github.com/hyperledger/fabric-x-common/tools/test"
 )
